S02P13C112-48 줄바꿈 적용

+++ back-end/src/main/java/edu/ssafy/boot/controller/UserController.java
@@ -185,9 +185,12 @@ public class UserController {
 
 		try {
 			user = ser.info(user_id);
+			String url = user.getProfile_url();
+			String imageName = url.substring(url.lastIndexOf("/")+1);
+			String ext = imageName.substring(imageName.lastIndexOf(".")+1);
 			Map<String, Object> map = new HashMap<String, Object>();
 			if (user != null) {
-				File file = new File(realPath + File.separator + user_id + ".jpg");
+				File file = new File(realPath + File.separator + imageName);
 				if (file.exists()) {
 
 					fis = new FileInputStream(file);
@@ -200,7 +203,7 @@ public class UserController {
 					}
 					byte[] fileArray = bos.toByteArray();
 					String imageString = new String(Base64.encodeBase64(fileArray));
-					String changeString = "data:image/jpg;base64, "+imageString;
+					String changeString = "data:image/"+ext+";base64, "+imageString;
 					user.setProfileImage(new ImageVo());
 					user.getProfileImage().setBase64(changeString);
 					user.getProfileImage().setFilter(user.getProfile_filter());
@@ -402,9 +405,12 @@ public class UserController {
 		if(user.getProfileImage() != null && user.getProfileImage().getBase64() != ""){
 
 			ImageVo image = user.getProfileImage();
+			String ext = image.getBase64().substring(image.getBase64().indexOf("/")+1, image.getBase64().indexOf(";"));
+			System.out.println(ext);
 				byte[] decode = Base64.decodeBase64(image.getBase64().substring(image.getBase64().lastIndexOf(",")));
-				String image_name = user.getUser_id() + ".jpg";
+				String image_name = user.getUser_id() + "."+ext;
 				String savePath = realPath+File.separator+image_name;
+				System.out.println(savePath);
 				String image_url = req.getScheme() + "://" + req.getServerName() + ":" + req.getServerPort() + path + "/" +image_name;
 				
 				File f = new File(savePath);
@@ -435,7 +441,13 @@ public class UserController {
 		String realPath = req.getServletContext().getRealPath(path);
 		
 		boolean isDelete = true;
-			String savePath = realPath+File.separator+user_id + ".jpg";
+		UserVo user = ser.info(user_id);
+			String url = user.getProfile_url();
+			String imageName = url.substring(url.lastIndexOf("/")+1);
+			if(imageName.equals("default.png")){
+				return isDelete;
+			}
+			String savePath = realPath+File.separator+imageName;
 			File file = new File(savePath);
 			if(file.exists()){
 				if(file.delete()){
